perf(eth): clean db files of ethereum listener
