+++ src/listener/eth/BlockchainState.ts
@@ -1,16 +1,16 @@
 import { localConfig as Config } from "./Config";
-import { blockInDB } from "./BlockInDB";
+import { blockInDB } from "./DB";
 
 export class BlockchainState {
-	getState(): Promise<any> {
-		return BlockchainState.getBlockState().then(([blockInChain, latestBlockInDB]) => {
-			blockInDB.setLastBlockNumber(blockInChain);
-		})
-	}
+    getState(): Promise<any> {
+        return BlockchainState.getBlockState().then(([blockInChain, latestBlockInDB]) => {
+            blockInDB.setLastBlockNumber(blockInChain);
+        })
+    }
 
-	static getBlockState(): Promise<any[]> {
-		const latestBlockOnChain = Config.web3.eth.getBlockNumber();
-		const latestBlockInDB = blockInDB.getBlockNumber();
-		return Promise.all([latestBlockOnChain, latestBlockInDB]);
-	}
+    static getBlockState(): Promise<any[]> {
+        const latestBlockOnChain = Config.web3.eth.getBlockNumber();
+        const latestBlockInDB = blockInDB.getBlockNumber();
+        return Promise.all([latestBlockOnChain, latestBlockInDB]);
+    }
 }
