+++ src/listener/eth/Starter.ts
@@ -1,8 +1,7 @@
 import { BlockchainState } from "./BlockchainState";
 import { EventParser } from "./EventParser";
-import { logInDB } from "./LogInDB";
+import { blockInDB, logInDB } from "./DB";
 import { localConfig } from "./Config";
-import { blockInDB } from "./BlockInDB";
 import { LogType } from "./types";
 
 const blockchainState = new BlockchainState();
@@ -11,11 +10,8 @@ const eventParser = new EventParser();
 export default class {
     start(config: any, callback: (tx: string, type: LogType) => void) {
         localConfig.setConfig(config);
-
         blockInDB.setParsedEventBlockNumber(localConfig.info.START_BLOCK_NUMBER);
-
         logInDB.setCallback(callback);
-
         blockchainState.getState().then(() => {
             eventParser.start(blockchainState);
         })
