+++ src/util/delay.ts
@@ -1,5 +1,4 @@
-
-export function setDelay(t: number): Promise<void> {
+export function delay(t: number): Promise<void> {
     return new Promise((resolve) => {
         setTimeout(resolve, t);
     });
