+++ src/util/index.ts
@@ -2,11 +2,13 @@ import { Config, TYPES_URL } from "./cfg";
 import { download } from "./download";
 import { log } from "./log";
 import { whereisPj } from "./pj";
+import { delay } from "./delay";
 import chalk from "chalk";
 
 // exports
 export {
     Config,
+    delay,
     chalk,
     download,
     log,
